efi-boot-generator: set automount expire for /boot
This should make the Linux vfat handling less fragile, by unmounting
the idle mount and causing to sync the entire file system to disk.

On my machines, The FAT file system at /boot regularly gets corrupted
with unclean shutdowns, rendering the machines unbootable. Either the
VFS write-back behaviour or the vfat driver itself is just not reliable
enough to be/stay mounted at every bootup.

Because of that, the automount expiry was on top of my personal
wish list. Thanks a lot Michael!