firewall-util: switch from iptables to nftables
nftables is the new packet classification framework that intends to replace
the existing {ip,ip6,arp,eb}_tables infrastructure.

The advantage over iptables in the case of systemd is that we can put all
systemd related chains and rules into a table of its own, so it doesn't
clash with other tools setting up firewall rules on the host.

The features of nftables needed by systemd such as masquerading support are
available since Linux kernel v3.18. iptables and nft cannot be used to
perform NAT at the same time, so the iptable_nat module must be unloaded in
order to make these features work.

The API has changed so that fw_add_masquerade() and fw_add_local_dnat()
return a uint64_t handle when entries are created. These handles can be used
to remove a rule again, which is what the new functions
fw_remove_masquerade() and fw_remove_local_dnat() do.