sd-network: ignore missing directory if networkd is disabled
sd_network_monitor is used by timesyncd and resolved. Both can be enabled
while networkd is disabled. /run/systemd/netif/links/ only exists if
networkd is enabled at build-time, so ignore the corresponding error if it
is disabled.
The user that owns /run/systemd/netif/links/ may not exist if networkd is
disabled, so always creating this directory is not possible.